refactor(Result): AcceptableError Type
